requirements.txt
streamlit
pandas
plotly
scikit-learn
joblib
numpy
kafka-python
faker